Rename source dir to server \ Update relevant test files which required source/server directorty path
