+++ tests/user.test.js
@@ -1,14 +1,14 @@
 const request = require('supertest')
 const jwt = require('jsonwebtoken')
 const mongoose = require('mongoose')
-const User = require('../source/models/user')
-const app = require('../source/itemizer-app')
+const User = require('../server/models/user')
+const app = require('../server/itemizer-app')
 
 
 const userOneId = new mongoose.Types.ObjectId()
 const userOne = {
     _id: userOneId,
-    name: 'Vald',
+    name: 'Vlad',
     email: 'vlad@example.com',
     password: 'ThatPasss77!',
     tokens: [
